Remove CouchTransaction
